+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Latest 
 
+### Bug Fixes
+- Fix proper handling of ClickHouseResult to release DB connection https://github.com/ClickHouse/clickhouse-java/issues/1538
+- Fix getting response input stream for binary formats.
+    - https://github.com/ClickHouse/clickhouse-java/issues/1494
+    - https://github.com/ClickHouse/clickhouse-java/issues/1567
+    - https://github.com/ClickHouse/clickhouse-java/issues/1475
+    - https://github.com/ClickHouse/clickhouse-java/issues/1376
+
 ## 0.6.0-patch5
 
 ### Bug Fixes